requirements.txt
pymongo==4.6.0
python-dotenv==1.0.0
google-generativeai==0.3.1
faker==19.13.0
tqdm==4.66.1
fastapi==0.108.0
uvicorn==0.25.0
streamlit==1.29.0
bokeh==3.3.0
pydantic==2.5.2
pandas==2.1.3
numpy==1.26.2
